requirements.txt
# Digital Twin - Core Dependencies
anthropic>=0.18.0
requests>=2.31.0
influxdb-client>=1.38.0
redis>=5.0.0
chromadb>=0.4.18
numpy>=1.24.0
pandas>=2.1.0
asyncio>=3.4.3
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
